Split metro_transit.rs into subdirectory module
Break the 883-line metro_transit.rs monolith into focused submodules:
- constants.rs (43 lines): all metro system constants
- types.rs (55 lines): MetroStation, MetroLine, MetroStats data types
- state.rs (316 lines): MetroTransitState resource, core logic, Saveable impl
- systems.rs (139 lines): ECS systems (stats, costs, land value boost)
- tests.rs (301 lines): all unit tests
- mod.rs (55 lines): re-exports, MetroTransitPlugin

All pub items re-exported from mod.rs so external code is unaffected.

Part of #1389
